cleanup profiles
