+++ app/frontend/src/utils/timeAgo.ts
@@ -1,6 +1,7 @@
 import moment from "moment";
 
 export function timeAgo(milliseconds: number) {
+  //the backend fuzzes times to 15 minute intervals
   if (new Date().getTime() - milliseconds < 15 * 60 * 1000) {
     return "less than 15 minutes ago";
   }
